% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTO.r
\name{HTOcutoff}
\alias{HTOcutoff}
\title{HTOcutoff}
\usage{
HTOcutoff(object = NULL, assay = "HTO", method = "CLR", min_limit = 1)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{the assay name of Hashtag signals, Default is "HTO"}

\item{method}{normalize method, could be "CLR" or 'log'. Default is "CLR"}

\item{min_limit}{min limit of cutoff, default is 1}
}
\description{
function to determine cutoff for HTO signals
}
