% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTO.r
\name{HTOIdAssign}
\alias{HTOIdAssign}
\title{HTOIdAssign}
\usage{
HTOIdAssign(data, cutoff)
}
\arguments{
\item{data}{nromalized HTO data}

\item{cutoff}{the cutoff values for HTO signals}
}
\description{
assign HTO groups to each cell according to their HTO signals
}
