% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTO.r
\name{HybridDemultiplexing}
\alias{HybridDemultiplexing}
\title{HybridDemultiplexing}
\usage{
HybridDemultiplexing(
  object = NULL,
  cellhashing_label = NULL,
  genotype_label = NULL,
  hto_names = NULL,
  c_threshold = 0.7,
  snp_doublet_label = "doublet",
  snp_unassign_label = "unassigned",
  cellhashing_doublet_label = "Doublet",
  cellhashing_negative_label = "Negative"
)
}
\arguments{
\item{object}{Seurat object}

\item{cellhashing_label}{label name for cell hashing in the meta data}

\item{genotype_label}{label name for snp-based method in the meta data}

\item{hto_names}{a vector of all HTO names}

\item{c_threshold}{threshold of convergence score, default is 0.7.}

\item{snp_doublet_label}{doublet label in SNP results.}

\item{snp_unassign_label}{unassigned label in SNP results.}

\item{cellhashing_doublet_label}{doublet label in cell hashing results.}

\item{cellhashing_negative_label}{negative label in cell hashing results.}
}
\description{
function for HybridDemultiplexing
}
