% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTO.r
\name{PlotHTO}
\alias{PlotHTO}
\title{PlotHTO}
\usage{
PlotHTO(
  object = NULL,
  assay = "HTO",
  method = "log",
  cutoff = NULL,
  xlim = NULL,
  plot_existing_cutoff = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{the assay name of Hashtag signals, Default is "HTO"}

\item{method}{normalize method, could be "CLR" or 'log'. Default is "log"}

\item{cutoff}{users can choose to display cutoff for each HTO signal by giving a data array}

\item{xlim}{max limit of x axis. Default is "NULL". Can be used to trim the density plot for better visualization of positive peaks and negative peaks}

\item{plot_existing_cutoff}{plot existing cutoff if the method fits}
}
\description{
function to plot histgram of each HTO signal (and their cutoff)
}
