% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTO.r
\name{HTOClassification}
\alias{HTOClassification}
\title{HTOClassification}
\usage{
HTOClassification(
  object = NULL,
  assay = "HTO",
  method = "CLR",
  specify_cutoff = NULL,
  min_limit = 1
)
}
\arguments{
\item{object}{Seurat object}

\item{assay}{the assay name of Hashtag signals, Default is "HTO"}

\item{method}{normalize method, could be "CLR" or 'log'. Default is "CLR"}

\item{specify_cutoff}{users can specify their own cutoffs for HTOs instead of letting program to determine them}

\item{min_limit}{the min limit of cutoff}
}
\description{
main function for HTO group Classification
}
